+++ R/draw_elev2D_OCN.R
@@ -6,12 +6,13 @@ draw_elev2D_OCN <- function(OCN,
                             thrADraw=0.002*OCN$FD$nNodes*OCN$cellsize^2,
                             riverColor="#00BFFF",
                             min_lwd=0.5,
-                            max_lwd=5){
+                            max_lwd=5,
+                            args_imagePlot=list()){
   
   if (!("Z" %in% names(OCN$FD))){
     stop('Missing fields in OCN. You should run landscape_OCN prior to draw_elev2D_OCN.')
   }
-
+  
   # plot elevation map
   #old.par <- par(no.readonly =TRUE) 
   #on.exit(par(old.par))
@@ -33,19 +34,30 @@ draw_elev2D_OCN <- function(OCN,
       Zmat[OCN$FD$toDEM] <- OCN$FD$Z
       Zmat <- Zmat[,seq(OCN$dimY,1,-1)]
       Zmat <- t(Zmat)
-      }
+    }
   } else {Zmat <- matrix(data=OCN$FD$Z,nrow=OCN$dimY,ncol=OCN$dimX)}
   
+  if(is.null(args_imagePlot$col)){args_imagePlot$col <- colPalette}
+  if(is.null(args_imagePlot$zlim)){args_imagePlot$zlim <- range(Zmat,na.rm=T)}
+  args_imagePlot$x <- seq(xllcorner,xllcorner+(OCN$dimX-1)*OCN$cellsize,OCN$cellsize)
+  args_imagePlot$y <- seq(yllcorner,yllcorner+(OCN$dimY-1)*OCN$cellsize,OCN$cellsize)
+  args_imagePlot$z <- t(Zmat)
+  if(is.null(args_imagePlot$xlab)){args_imagePlot$xlab <- " "} 
+  if(is.null(args_imagePlot$ylab)){args_imagePlot$ylab <- " "}
+  if(is.null(args_imagePlot$axes)){args_imagePlot$axes <- FALSE} 
+  if(is.null(args_imagePlot$asp)){args_imagePlot$asp <- 1}
+  if(is.null(args_imagePlot$xlim)){args_imagePlot$xlim=range(OCN$FD$X)}
+  if(is.null(args_imagePlot$ylim)){args_imagePlot$ylim=range(OCN$FD$Y)}
+  
   if (addLegend==TRUE){
-  image.plot(Xvec, Yvec,
-             t(Zmat),col=colPalette,xlab=" ",ylab=" ",axes=FALSE,asp=1,
-             xlim=range(OCN$FD$X), ylim=range(OCN$FD$Y))
-  } else {
-    image(seq(xllcorner,xllcorner+(OCN$dimX-1)*OCN$cellsize,OCN$cellsize),
-          seq(yllcorner,yllcorner+(OCN$dimY-1)*OCN$cellsize,OCN$cellsize),
-               t(Zmat),col=colPalette,xlab=" ",ylab=" ",axes=FALSE,asp=1,
-          xlim=range(OCN$FD$X), ylim=range(OCN$FD$Y))
-  }
+    do.call(imagePlot, args_imagePlot)
+  } else {do.call(image, args_imagePlot)}
+  
+  # image(seq(xllcorner,xllcorner+(OCN$dimX-1)*OCN$cellsize,OCN$cellsize),
+  #       seq(yllcorner,yllcorner+(OCN$dimY-1)*OCN$cellsize,OCN$cellsize),
+  #       t(Zmat),col=colPalette,xlab=" ",ylab=" ",axes=FALSE,asp=1,
+  #       xlim=range(OCN$FD$X), ylim=range(OCN$FD$Y))
+  
   
   if (drawRiver==TRUE){
     AvailableNodes <- setdiff(1:OCN$FD$nNodes,OCN$FD$outlet)